requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2
asyncpg
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
authlib
python-multipart
itsdangerous
httpx
email-validator
